requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pymongo
google-genai
httpx
pydantic
